requirements.txt
pytest==7.4.0
selenium==4.11.0
allure-pytest==2.13.0
webdriver-manager==3.8.6